% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{dtFetchUsTreasuryYieldCurveData}
\alias{dtFetchUsTreasuryYieldCurveData}
\title{Fetch the US Treasury Data}
\usage{
dtFetchUsTreasuryYieldCurveData(
  cDataType = "YieldCurveRates",
  dateStartDate,
  dateEndDate
)
}
\arguments{
\item{cDataType}{- character scalar, one of the following three values:
"YieldCurveRates", "BillRates", "LongTermRates"}

\item{dateStartDate}{- class 'Date' scalar}

\item{dateEndDate}{- class 'Date' scalar; function checks if the dateEndDate
precedes dateStartDate and throws an error should this be the case}
}
\description{
Fetching nominal interest rates published by US Treasury
}
\details{
This function fetches the historical time series of nominal interest rates
as published by US Treasury at the site:
As only the nominal interest rates are handled, the following data types
are available (cf. parameter cDataType): 1) Yield Curve Rates,
2) Bill Rates and 3) Long-Term Rates
}
